ai_interpreter/core/interpreter.py
def generate_command(input_text):
    """
    Generate a command from user input.
    Implement your command generation logic here.
    """

    return input_text


def execute_command(command):
    """
    Execute the generated command.
    Implement your command execution logic here.
    """
    if command == "exit":
        return "Goodbye!"
    elif command == "help":
        return "Available commands: run, exit, help, info"
    elif command == "info":
        return "This is an AI interpreter."
    elif command == "run":
        return "Running the 'run' command."
    else:
        return f"Unknown command: {command}. Type 'help' for available commands."
